+++ packages/react-app/docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.0.0-SNAPSHOT
 
+1. Fix CI/CD for preview and production
+2. Add bounty filters
+
+## 1.0.0-beta2 (2021-08-26)
+
 1. Use QA specific environment variables
 2. Add metadata to display preview in discord link
 3. Adjust hashID for discordjs webhook message